server-pae/requirements.txt
django==4.0.4
djangorestframework
psycopg2-binary
django-cors-headers